% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.decp_result}
\alias{plot.decp_result}
\title{Plot method for decp_result}
\usage{
\method{plot}{decp_result}(x, ...)
}
\arguments{
\item{x}{An object of class 'decp_result'}

\item{...}{Additional arguments passed to the plotting function}
}
\description{
Plot method for decp_result
}
