% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.mle_change_point_result}
\alias{plot.mle_change_point_result}
\title{Plot method for mle_change_point_result}
\usage{
\method{plot}{mle_change_point_result}(x, ...)
}
\arguments{
\item{x}{An object of class 'mle_change_point_result'}

\item{...}{Additional arguments passed to the plotting function}
}
\description{
Plot method for mle_change_point_result
}
