% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_change_points}
\alias{plot_change_points}
\title{Plot Change Points}
\arguments{
\item{estimation_results}{A list of estimation results.}

\item{ordered_change_points}{A vector of ordered change points.}
}
\value{
A ggplot object.
}
\description{
This function creates a density plot of change points from estimation results.
}
\examples{
# Example usage
estimation_results <- list(estimation_vecRW_01 = rnorm(100), estimation_vecRW_02 = rnorm(100))
ordered_change_points <- c(50, 150)
plot_change_points(estimation_results, ordered_change_points)
}
