% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decp.R
\name{print.decp_result}
\alias{print.decp_result}
\title{Print method for decp_result}
\usage{
\method{print}{decp_result}(x, ...)
}
\arguments{
\item{x}{An object of class 'decp_result'}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for decp_result
}
