% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_change_point.R
\name{print.mle_change_point_result}
\alias{print.mle_change_point_result}
\title{Print method for 'mle_change_point_result' class}
\usage{
\method{print}{mle_change_point_result}(x, ...)
}
\arguments{
\item{x}{An object of class 'mle_change_point_result'.}

\item{...}{Additional arguments (not used).}
}
\description{
Print method for 'mle_change_point_result' class
}
