% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decp.R
\name{summary.decp_result}
\alias{summary.decp_result}
\title{Summary method for decp_result}
\usage{
\method{summary}{decp_result}(object, ...)
}
\arguments{
\item{object}{An object of class 'decp_result'}

\item{...}{Additional arguments (not used)}
}
\description{
Summary method for decp_result
}
